+++ src/includes.hpp
@@ -115,6 +115,9 @@ public:
     bool seedEnabled = false;
     bool clickbotEnabled = false;
     bool frameLabel = false;
+    bool trajectoryBothSides = false;
+    bool p2mirror = false;
+    bool lockDelta = false;
 
     bool ignoreStopDashing[2] = { false, false };
     bool addSideHoldingMembers[2] = { false, false };