+++ src/ui/render_settings_layer.hpp
@@ -6,6 +6,11 @@ class RenderSettingsLayer : public geode::Popup<>, public TextInputDelegate {
 	
 public:
 
+	Slider* sfxSlider = nullptr;
+	Slider* musicSlider = nullptr;
+	TextInput* fadeInInput = nullptr;
+	TextInput* fadeOutInput = nullptr;
+
 	CCTextInputNode* argsInput = nullptr;
 	CCTextInputNode* audioArgsInput = nullptr;
 	CCTextInputNode* secondsInput = nullptr;
@@ -22,7 +27,7 @@ private:
 
 public:
 
-	STATIC_CREATE(RenderSettingsLayer, 268, 277)
+	STATIC_CREATE(RenderSettingsLayer, 396, 277)
 	
 	void open(CCObject*) {
 		create()->show();
@@ -33,6 +38,8 @@ public:
 	}
 
 	void textChanged(CCTextInputNode* node) override;
+
+	void onSlider(CCObject*);
 
 	void onDefaults(CCObject*);
 };